:fire: remove apps utils
