+++ lib/applications/index.js
@@ -8,10 +8,8 @@
 
 const model = require('./model');
 const routes = require('./routes');
-const utils = require('./utils');
 
 module.exports = {
   model,
   routes,
-  utils,
 };
